+++ other/mq_client_example/example/my_record.go
